tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe-allowfullscreen.html

<!doctype html>
<meta charset=utf-8>
<title>Check how allowfullscreen affects fullscreen enabled flag</title>
<link rel="author" title="Xidorn Quan" href="https://www.upsuper.org">
<link rel="author" title="Mozilla" href="https://www.mozilla.org">
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#initialise-the-document-object">
<link rel="help" href="https://fullscreen.spec.whatwg.org/#fullscreen-enabled-flag">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>
<script>
  // This returns a data URL (cross-origin with the containing document) which
  // advances a counter, and reports the counter value together with the
  // document's fullscreenEnabled state, every time it receives a postMessage.
  function getSourceForCrossOriginPage(initial_count) {
    var page_contents = "<html><body><script>var count="+initial_count+";window.addEventListener('message',function(){parent.postMessage({'count':count++,'fullscreenEnabled':document.fullscreenEnabled},'*');});</scr"+"ipt></body></html>";
    return "data:text/html;base64,"+btoa(page_contents);
  }

  async_test(function(t) {
    var iframe = document.createElement("iframe");
    iframe.src = "support/blank.htm";
    var eventWatcher = new EventWatcher(t, iframe, "load");
    document.body.appendChild(iframe);
    t.add_cleanup(function() {
      document.body.removeChild(iframe);
    });

    assert_true(document.fullscreenEnabled, "Top level document has fullscreen enabled flag set");
    eventWatcher.wait_for("load").then(t.step_func_done(function() {
      assert_true(iframe.contentDocument.fullscreenEnabled, "Document inside same-origin iframe without allowfullscreen attribute should still have fullscreen enabled flag set");
    }));
  }, "iframe-same-origin-allowfullscreen");

  async_test(function(t) {
    var iframe = document.createElement("iframe");
    iframe.src = getSourceForCrossOriginPage(0);

    iframe.addEventListener('load', function() {
      // Request the fullscreenEnabled state whenever the frame loads
      iframe.contentWindow.postMessage(true,"*");
    });

    window.addEventListener('message', this.step_func(function(msg) {
      if (msg.data.count == 0) {
        assert_false(msg.data.fullscreenEnabled, "Document inside cross-origin iframe without allowfullscreen attribute should not have fullscreen enabled flag set");
        iframe.setAttribute("allowfullscreen", "");
        iframe.contentWindow.postMessage(true,"*"); // Request state again
      } else if (msg.data.count == 1) {
        assert_false(msg.data.fullscreenEnabled, "Fullscreen should be denied when allowfullscreen attribute is added, before reload");
        iframe.src = getSourceForCrossOriginPage(2); // Reload the frame
      } else if (msg.data.count == 2) {
        assert_true(msg.data.fullscreenEnabled, "Fullscreen should be allowed when allowfullscreen attribute is added, after reload");
        iframe.removeAttribute("allowfullscreen");
        iframe.contentWindow.postMessage(true,"*"); // Request state again
      } else if (msg.data.count == 3) {
        assert_true(msg.data.fullscreenEnabled, "Fullscreen should be allowed when allowfullscreen attribute is removed, before reload");
        iframe.src = getSourceForCrossOriginPage(4); // Reload the frame
      } else if (msg.data.count == 4) {
        assert_false(msg.data.fullscreenEnabled, "Fullscreen should be denied when allowfullscreen attribute is removed, after reload");
        t.done();
      }
    }));

    document.body.appendChild(iframe);
    t.add_cleanup(function() {
      document.body.removeChild(iframe);
    });
  }, "iframe-cross-origin-allowfullscreen");

  /* Fullscreen enabled flag with about:blank */

  function test_allowfullscreen_noload(setup_iframe, check) {
    var iframe = document.createElement("iframe");
    setup_iframe(iframe);
    document.body.appendChild(iframe);
    check(iframe.contentDocument);
    document.body.removeChild(iframe);
  }

  test(function() {
    test_allowfullscreen_noload(function() {}, function(doc) {
      assert_true(doc.fullscreenEnabled, "Fullscreen should still be enabled in same-origin document without allowfullscreen attribute");
    });
  }, "iframe-noload-noallowfullscreen");

  test(function() {
    test_allowfullscreen_noload(function(iframe) {
      iframe.setAttribute("allowfullscreen", true);
    }, function(doc) {
      assert_true(doc.fullscreenEnabled, "Fullscreen should be enabled with allowfullscreen attribute");
    });
  }, "iframe-noload-allowfullscreen");
</script>
